% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countops_detail}
\alias{countops_detail}
\title{Detailed CountOps data.}
\format{A data.frame with 2,094,354 rows and columns
\itemize{
\item Hour - Hour of the day, 0-23
\item Runway
\item Operation - Arrival, Departure or Overflight
\item Equipment - CountOps code for the type of aircraft
\item Count - Number of operations
\item Date
\item Category - Wide body, single aisle, etc
\item Jet - true / false
}}
\source{
FAA CountOps data via FOIA request.
}
\usage{
countops_detail
}
\description{
Detailed CountOps data.
}
\keyword{datasets}
