% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countops_hourly}
\alias{countops_hourly}
\title{CountOps data summarized by hour, runway and equipment category.}
\format{A data.frame with 908,613 rows and columns
\itemize{
\item Date
\item Hour - Hour of the day, 0-23
\item Runway
\item Operation - Arrival, Departure or Overflight
\item Category - Wide body, single aisle, etc
\item Count - Number of operations
\item Month - Date of the first of the month containing the event
\item Jet - true / false
}}
\source{
FAA CountOps data via FOIA request.
}
\usage{
countops_hourly
}
\description{
CountOps data summarized by hour, runway and equipment category.
}
\keyword{datasets}
