% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countops_hourly_jets}
\alias{countops_hourly_jets}
\title{CountOps data, jets only, summarized by hour and runway.}
\format{A data.frame with 263,490 rows and columns
\itemize{
\item Date
\item Month - Date of the first of the month containing the event
\item Hour - Hour of the day, 0-23
\item Runway
\item Operation - Arrival, Departure or Overflight
\item Count - Number of operations
}}
\source{
FAA CountOps data via FOIA request.
}
\usage{
countops_hourly_jets
}
\description{
CountOps data, jets only, summarized by hour and runway.
}
\keyword{datasets}
